requirements.txt
requests
beautifulsoup4
supabase
python-dotenv